# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Clip Studio API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export type ClipMode = (typeof ClipMode)[keyof typeof ClipMode];

export const ClipMode = {
  edited: "edited",
  raw: "raw",
} as const;

export type ClipSourceType =
  (typeof ClipSourceType)[keyof typeof ClipSourceType];

export const ClipSourceType = {
  youtube: "youtube",
  local: "local",
} as const;

export type ClipStatus = (typeof ClipStatus)[keyof typeof ClipStatus];

export const ClipStatus = {
  pending: "pending",
  processing: "processing",
  done: "done",
  error: "error",
} as const;

export interface Clip {
  id: number;
  /** @nullable */
  youtubeUrl: string | null;
  startTime: string;
  endTime: string;
  headline: string;
  mode: ClipMode;
  sourceType: ClipSourceType;
  /** @nullable */
  localFileName: string | null;
  status: ClipStatus;
  /**
   * @minimum 0
   * @maximum 100
   */
  progress: number;
  /** @nullable */
  errorMessage: string | null;
  /** @nullable */
  outputFilename: string | null;
  createdAt: string;
  updatedAt: string;
  sourceChannel: string;
}

export type CreateClipBodyMode =
  (typeof CreateClipBodyMode)[keyof typeof CreateClipBodyMode];

export const CreateClipBodyMode = {
  edited: "edited",
  raw: "raw",
} as const;

export interface CreateClipBody {
  youtubeUrl: string;
  /** Start timestamp in HH:MM:SS format */
  startTime: string;
  /** End timestamp in HH:MM:SS format */
  endTime: string;
  headline?: string;
  mode?: CreateClipBodyMode;
  sourceChannel?: string;
}

export interface ClipStats {
  total: number;
  pending: number;
  processing: number;
  done: number;
  error: number;
}