# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Clip Studio API specification
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * Returns server health status
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * Returns all clip jobs, newest first
 * @summary List all clips
 */
export const listClipsResponseProgressMin = 0;
export const listClipsResponseProgressMax = 100;

export const ListClipsResponseItem = zod.object({
  id: zod.number(),
  youtubeUrl: zod.string().nullable(),
  startTime: zod.string(),
  endTime: zod.string(),
  headline: zod.string(),
  mode: zod.enum(["edited", "raw"]),
  sourceType: zod.enum(["youtube", "local"]),
  localFileName: zod.string().nullable(),
  frameStyle: zod.enum(["standard", "immersive"]),
  status: zod.enum(["pending", "processing", "done", "error"]),
  progress: zod
    .number()
    .min(listClipsResponseProgressMin)
    .max(listClipsResponseProgressMax),
  errorMessage: zod.string().nullable(),
  outputFilename: zod.string().nullable(),
  createdAt: zod.coerce.date(),
  updatedAt: zod.coerce.date(),
  sourceChannel: zod.string(),
  captionsEnabled: zod.boolean(),
  transcript: zod.string().nullable(),
});
export const ListClipsResponse = zod.array(ListClipsResponseItem);

/**
 * Submits a new clip processing job
 * @summary Create a clip job
 */
export const createClipBodyHeadlineDefault = ``;
export const createClipBodyModeDefault = `edited`;
export const createClipBodyFrameStyleDefault = `immersive`;

export const CreateClipBody = zod.object({
  youtubeUrl: zod.string(),
  startTime: zod.string().describe("Start timestamp in HH:MM:SS format"),
  endTime: zod.string().describe("End timestamp in HH:MM:SS format"),
  headline: zod.string().default(createClipBodyHeadlineDefault),
  mode: zod.enum(["edited", "raw"]).default(createClipBodyModeDefault),
  frameStyle: zod.enum(["standard", "immersive"]).default(createClipBodyFrameStyleDefault),
  sourceChannel: zod.string().default(""),
  captionsEnabled: zod.boolean().default(true),
});

/**
 * @summary Get a clip by ID
 */
export const GetClipParams = zod.object({
  id: zod.coerce.number(),
});

export const getClipResponseProgressMin = 0;
export const getClipResponseProgressMax = 100;

export const GetClipResponse = zod.object({
  id: zod.number(),
  youtubeUrl: zod.string().nullable(),
  startTime: zod.string(),
  endTime: zod.string(),
  headline: zod.string(),
  mode: zod.enum(["edited", "raw"]),
  sourceType: zod.enum(["youtube", "local"]),
  localFileName: zod.string().nullable(),
  frameStyle: zod.enum(["standard", "immersive"]),
  status: zod.enum(["pending", "processing", "done", "error"]),
  progress: zod
    .number()
    .min(getClipResponseProgressMin)
    .max(getClipResponseProgressMax),
  errorMessage: zod.string().nullable(),
  outputFilename: zod.string().nullable(),
  createdAt: zod.coerce.date(),
  updatedAt: zod.coerce.date(),
  sourceChannel: zod.string(),
  captionsEnabled: zod.boolean(),
  transcript: zod.string().nullable(),
});

/**
 * @summary Delete a clip
 */
export const DeleteClipParams = zod.object({
  id: zod.coerce.number(),
});

/**
 * @summary Retry a failed clip
 */
export const RetryClipParams = zod.object({
  id: zod.coerce.number(),
});

export const retryClipResponseProgressMin = 0;
export const retryClipResponseProgressMax = 100;

export const RetryClipResponse = zod.object({
  id: zod.number(),
  youtubeUrl: zod.string().nullable(),
  startTime: zod.string(),
  endTime: zod.string(),
  headline: zod.string(),
  mode: zod.enum(["edited", "raw"]),
  sourceType: zod.enum(["youtube", "local"]),
  localFileName: zod.string().nullable(),
  frameStyle: zod.enum(["standard", "immersive"]),
  status: zod.enum(["pending", "processing", "done", "error"]),
  progress: zod
    .number()
    .min(retryClipResponseProgressMin)
    .max(retryClipResponseProgressMax),
  errorMessage: zod.string().nullable(),
  outputFilename: zod.string().nullable(),
  createdAt: zod.coerce.date(),
  updatedAt: zod.coerce.date(),
  captionsEnabled: zod.boolean(),
  transcript: zod.string().nullable(),
});

/**
 * Returns aggregate stats about clips
 * @summary Get clip processing statistics
 */
export const GetClipStatsResponse = zod.object({
  total: zod.number(),
  pending: zod.number(),
  processing: zod.number(),
  done: zod.number(),
  error: zod.number(),
});